Create MeciaCode.ino
The implementation of the Arduino-based Smart Plant Monitoring System will be done on the Blynk platform in order to observe or manage the plant remotely according to the measurement of environmental parameters such as temperature, humidity, and soil moisture. It sends notifications to the user in case of low soil moisture or extremely high temperatures of the plant.

Key Features:
The DHT11 Sensor is responsible for measuring the humidity and ambient temperature, further sending data to the Blynk application.
Dallas Temperature Sensor: It measures the temperature of the soil and presents real-time data to the user. Soil Moisture Sensor: This reads the moisture level of the soil. In case the moisture goes below the threshold value of 50%, the system will trigger a Blynk event to notify the user to switch on the water pump. Water Pump Control: This can be interfaced such that the irrigation of the plants can either be scheduled or done manually via the Blynk app with regards to the moisture values of the soil.
Blynk Notifications: The system sends log events through the Blynk platform. It notifies the user in the following cases: Soil moisture less than a prescribed value, Temperature rises above a set threshold value, such as 25°C. Code Description: System reads humidity and temperature from DHT11 every second and refreshes on Blynk virtual pins. Every 2 seconds, it measures the soil moisture and temperature. If the values cross the predefined thresholds, notification events are triggered.
All sensor data will be sent to the Blynk app for remote monitoring of conditions and allow the user to take further necessary actions from a distance, like turning on the water pump.
This project is ideal for home gardeners and plant enthusiasts who want a seamless, automated plant monitoring system by means of IoT.